// src/api/gen/models/ExercisesSearchRequest.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { ExerciseFilter } from "./ExerciseFilter";
/**
 * Запрос на поиск упражнений
 */
export type ExercisesSearchRequest = {
  filter?: ExerciseFilter;
  page?: number;
  size?: number;
};